# parser: accelerate DecodeLayers