Most of frequently used (and natively supported) LayerType values are not very big, so we may attempt to use array to search `DecodingLayer`, fallback to map only if `LayerType` is too big. Similar approach as in `RegisterLayerType`. I chose 256 as a decent threshold.

Results on Xeon(R) CPU E5-2620:
```
benchmark                                      old ns/op     new ns/op     delta
BenchmarkDecodingLayerParserIgnorePanic-32     227           191           -15.86%
BenchmarkDecodingLayerParserHandlePanic-32     291           227           -21.99%
```