@yerden 2019-05-30
And one more thing. I don't want to overgeneralize (and please correct me if i'm wrong) but I suspect that most common usage pattern here suggests not so many DecodingLayer-s in a single DecodingLayerParser. Most of the time, up to ten layers is enough.

If that's true, naive linear search of DecodingLayer beats `map`, my laptop (Core i5-3320M) tests:
```
BenchmarkSearchMap3-4       	30000000	        50.4 ns/op
BenchmarkSearchMap5-4       	30000000	        56.2 ns/op
BenchmarkSearchMap10-4      	20000000	        73.2 ns/op

BenchmarkSearchArray3-4     	30000000	        40.2 ns/op
BenchmarkSearchArray5-4     	30000000	        39.3 ns/op
BenchmarkSearchArray10-4    	30000000	        39.5 ns/op

BenchmarkSearchLinear3-4    	30000000	        45.2 ns/op
BenchmarkSearchLinear5-4    	30000000	        50.0 ns/op
BenchmarkSearchLinear10-4   	30000000	        52.8 ns/op
```

That's a definite win if some custom LayerType values are in use, so you don't want to use array indexing (too big values).

And, in case of a simple Eth->IPv4->TCP packet, about 60-70% of DecodeLayers() call amounts to searching for an appropriate DecodingLayer rather than layer decoding itself.