@yerden 2019-05-30
I will investigate other ways to boost performance and will return later.