@yerden 2019-05-30
Well, other ideas came up to me, like:

1. `DecodingLayerParser` doesn't consider level of a layer. For example, LayerTypeEthernet may encounter at the beginning only, rather than in middle of a packet, so when looking for the first level decoder, `layers.Ethernet` should be the primary suspect to test against.

2. As I said earlier, in case we have only few decoders, linear search is preferable to `map`.

3. Proposed approach is good when using gopacket-owned LayerType values but it doesn't work when working with custom LayerType values. Maybe we should implement different array values >1000.

4. Maybe we could provide end-user with a tool to supply his own map implementation for LayerType->DecodingLayer lookup. This would require to break API.

So... I though I should take a break, test these ideas and revisit later. Or, alternatively, I can simply reopen this PR, allow it to merge and work on it for next PR. Any ideas?